ml_model/autoencoder/requirements.txt
scikit-learn==0.23.1 # 6.8 MB
numpy==1.19.0 # 14.6 MB
pandas==1.0.5 # 10.1 MB
tensorboard==2.1.1 # 3.8 MB
tensorflow-cpu@ https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow_cpu-2.1.0-cp36-cp36m-manylinux2010_x86_64.whl
tensorflow-estimator==2.1.0

# 수정 필요